Use absolute directory of Outpost.exe for calls to access/modify outpost2.ini
Previously we were using the current working directory to access outpost2.ini. I believe the working directory is being changed to the directory of the source code when attaching the debugger and attempting to debug. This is causing problems with applying the addons netfix and nethelper.

Originally, when using a fresh copy of outpost.ini from the repository, NetFix would load on the first attempt. On subsequent debug sessions, NetHelper would not load. Also, a file named outpost2.ini with nothing inside would be created in the source code directory. Hooman witnessed this with me on our last pair programming session, but we got it to work with some fiddling, so we moved on.

Now, whenever op2ext's dllmain function is called, the working directory is manually set to the location of the Outpost 2 executable. This fixed all issues with NetFix. I also set all calls inside op2ext attempting to access/modify outpost2.ini to the proper absolute directory instead of relying on relative directories.

I think it would be best not to rely on the working directory for any calls.

Unfortunately, NetHelper.dll is currently throwing an exception after exiting out of Outpost 2. This occurs before DllMain is called in op2ext.h. My hunch is this may be related to NetHelper attempting to use the working directory set to the wrong location, but I haven't examined anything to verify that.

=== Exception Details ===

Exception thrown at 0x25005E9C (NetHelper.dll) in Outpost2.exe: 0xC0000005: Access violation reading location 0x000000B8.

DISASSEMBLY INFO
25005E92  call        25005A0E
25005E97  mov         ebx,eax
25005E99  mov         eax,dword ptr [edi+10h]
25005E9C  mov         esi,dword ptr [eax+ebx*8]
25005E9F  push        ebx
25005EA0  lea         eax,[ebp-8]
25005EA3  push        eax
25005EA4  mov         ecx,edi
25005EA6  mov         dword ptr [ebp-4],esi
25005EA9  call        25005D48

CALL STACK
>	NetHelper.dll!25005e9c()	Unknown	No symbols loaded.
 	[Frames below may be incorrect and/or missing, no symbols loaded for NetHelper.dll]		Annotated Frame
 	NetHelper.dll!250066a9()	Unknown	No symbols loaded.
 	NetHelper.dll!250068d4()	Unknown	No symbols loaded.
 	NetHelper.dll!25006a2c()	Unknown	No symbols loaded.
 	NetHelper.dll!25006aae()	Unknown	No symbols loaded.
 	NetHelper.dll!250069fa()	Unknown	No symbols loaded.
 	NetHelper.dll!250050d4()	Unknown	No symbols loaded.
 	NetHelper.dll!25005177()	Unknown	No symbols loaded.
 	NetHelper.dll!25005220()	Unknown	No symbols loaded.
 	NetHelper.dll!250053e8()	Unknown	No symbols loaded.
 	NetHelper.dll!25005521()	Unknown	No symbols loaded.
 	NetHelper.dll!2500767b()	Unknown	No symbols loaded.
 	NetHelper.dll!25007797()	Unknown	No symbols loaded.
 	NetHelper.dll!25007814()	Unknown	No symbols loaded.
 	[External Code]		Annotated Frame
 	Outpost2.exe!004c781f()	Unknown	No symbols loaded.
 	Outpost2.exe!004c772e()	Unknown	No symbols loaded.
 	Outpost2.exe!004c4261()	Unknown	No symbols loaded.
 	[External Code]		Annotated Frame