Update log time format for consistency
The following shortcuts are used for ISO 8601 date format
%F = %Y-%m-%d
%T = %H:%M:%S

This is in contrast to
%X = writes localized time representation (locale dependent)

At the heart of this may be who is reading the log files. Although it
may make sense to use locale dependent output for end users, the log
files are not typically read by end users. Perhaps we should prioritize
keeping the log output readable for developers, or for machine parsing.
Using a non-locale dependent format may help with that.

Finally, the ISO date format should still be readable to end users.