Make GetSystemDateTime a free standing function
Private static methods are often a bit silly. You can't access them from
outside the class, and they can't access class member variables. This
effectively makes most such methods independent from the class.

An exception to the above is if the function needs access to private
class members. This might be to access private static variables, such as
instance counters, or singleton instance variables. It may also be
needed to call private methods on a class instance. The class instance
obviously can't come from a "this" pointer for a static method, so this
assumes the private static method either creates a class instances, or
gets one from someone else, such as a parameter, a global, or a static
variable.

None of the above exceptions apply here.