Extract CountFilesByTypeInDirectory to own file
This is mainly to segregate uses of the C++17 filesystem header. On
Linux, filesystem code can be checked with GCC or Clang, but not Mingw.
Windows code can be checked with Mingw, but not GCC or Clang. By
segregating filesystem code, it makes it easier to do test compiles of
individual files, or otherwise use the Linux compilers to scan the code
for errors.

This method also seems generally useful in it's own right. Taken a step
further, perhaps it can be converted into a custom iterator class, used
to iterate over all files in a directory with a given extension.