Replace `operator[]` with `.end()`
This provides an alternate way to get the addresses past the end of the
array, without using `operator[]`. Use of `operator[]` past the end of
the array may constitute a de-reference past the end, which would be
undefined behaviour. In particular, `p[size]` is equivalent to `*(p +
size)`. Taking the address of this results in the expression `&(*(p +
size))`. The debate is then does `&*p` constitute taking the address of
a de-referenced pointer, or is it considered a no-op. The C++ standard
does not seem to be clear on this point.

The C99 standard states such expressions are safe for built in arrays.
However, we are using C++ here and vectors, so the C99 standard does not
apply.