Convert lpMsgBuf to LPTSTR
This requires less casting. Converted the one needed C-syle cast to a
C++-style `reinterpret_cast`.

Normally parameter 5 to `FormatMessage` takes a pointer to a string
`LPTSTR`. However, when `FORMAT_MESSAGE_ALLOCATE_BUFFER` is specified,
it instead takes a pointer to a pointer to a string `LPTSTR*`. This
enables it to write a pointer to the allocated buffer to the parameter
as an out value (with a slightly different type). This is why the
`reinterpret_cast` is needed.