Unit tests depend on compiled code not full DLL
This effectively uses the compiled object files like a library, which it
can static link to, rather than trying to do a full DLL build, and link
to it at runtime.

The full DLL build is currently blocked on Linux due to name mangling
differences with a couple of Outpost2DLL imports. It's also blocked by
the use of `_ReturnAddress`.