Add `#pragma intrinsic(_ReturnAddress)`
This doesn't seem to be required, but is included in the Microsoft
documentation. Other sources suggest using it, but don't give a clear
reason why.

It may be because some intrinsics can be implemented as either a
callable method, or inlined. For `_ReturnAddress`, it would not behave
correctly if it was a callable method. I suspect the compiler only
allows this one to be expanded inline, so the `#pragma` may just be a
bit of paranoia.