Change default target to static library
Ideally, the DLL should be the final target. The static library should
be a dependent target.

As the DLL can not currently be built due to name mangling errors, avoid
setting it as the final target of the default rule. This avoids spurious
error messages about expected failure.

For now, to keep things simple, just try to build the static library.
That will provide error checking and unit testing ability. Coverage
should be reasonably complete, as most code is contained in the static
library portion. Eventually, we should also try to at least compile the
DLL source, even if it can not be linked. That will provide for maximum
testability, despite the link problems.