Update to use system Google Test include
Making a local copy of the include folder is no longer needed for cross
platform compiles. The new package structure supports this use directly.

The Google Test package has internal includes of the form `#include
<gtest/gtest.h>`, and also encourages such use with it's documentation.
That meant the parent folder needed to be part of the search path.

The previous install location was `/usr/include/gtest/`. This meant
adding `/usr/include/` to the search path. Unfortunately that is the
Linux system include path, which is not appropriate to add during a
cross compilation for Windows. In particular, many of the files in
`/usr/include/` would conflict with the Mingw versions used for
generating a Windows binary.

The new package places include files at:
`/usr/src/googletest/googletest/include/gtest/`
Hence, the level before can simply be included directly, as it contains
nothing else except for Google Test.