Adjust base development tools to install
The GCC compiler is used almost universally. Even if Clang or Mingw are
used, the GCC compiler will still be needed as a dependency of those
other compilers. As such GCC and associated build tools (such as make,
static library archivers, and linkers) make for a good shared base
image.

Add cmake to that list, as it's farily small, and almost universally
used. In particular, cmake is needed for a source install of Google
Test, which almost all OPU projects will make use of.