Remove accidentally committed test line
Previously I was checking the returned path against the current working
directory. Often this matches the exe path, but not when starting the
exe with a remote path, such as is done in the testing code. The other
reason was to get a dump of the current exe path, which is better
handled with the `<<` operator on the `EXPECT_*` tests.