Move "@" out of variable and into receipe
The decision as to print the command or not before executing it should
be decided in the receipe, not in the variable.

This would be relevant if the variable were to be used as a command
component, chained together into a compound command. An example might be
to run both commands in the same shell instance by combining them with
`&&`, `||`, or `;`. In that context the `@` would be passed to the
shell, where it is not understood and would cause an error.