Add intermediate target to projects
This allows a project to be compiled without the linking step. This can
be important for projects that are not able to fully compile with an
unsupported toolchain because of failures in the link step. By avoiding
the link step, and hence the failure, the toolchain is still able to
process the source file and report warnings or errors, but not cause
failures for a Continuous Integration build. This allows using
unsupported toolchains for error checking and linting support, without
having to fully support a complete build with them.