Update third parameter to `SourceFindPattern`
This parameter can still take a folder, which will be recursively
searched. That is the old behaviour.

New behaviour includes being able to set search pattern, which might
include wildcards for the filename portion. This is helpful for projects
that contain source code in the root of the repository, and also contain
dependencies in sub-folders.

Example:
./  # Search root of repository recursively (may include subprojects)
./*.cpp  # Search root of repository (but no recursive search)