Split Project macro into Variables and Targets
The high level `DefineCppProject` still exists and works as before.
However, it is now implemented in terms of two sub-parts,
`DefineCppProjectVariables` and `DefineCppProjectTargets`. This
breakdown allows for better debugging, as variables can be defined and
inspected before Make tries to process the rules that depend on them,
which could result in an error which aborts the Make run. It also means
you can dump the contents of one of the component variables with `$(info
$(variable))` to get more targetted output with less noise.