Add offset to account for `CALL` opcode
This is probably better than offsetting all the addresses by 1 manually.
When writing future patches, it may be easiest to copy+paste from an
OllyDbg reference search, which will show addresses of the `CALL`
instructions, rather than the relative offset field within the `CALL`
instruction.

This also means the method can verify the address is of a `CALL`
instruction.