Split `GetCommandLineArguments` to new file
This method is a bit unlike the other parsing methods. For one, it's
quite general, and usable in many contexts. For another, it's the only
method that actually touches the Windows API. By splitting it into a new
file, we can limit the code that needs to include `<windows.h>`. By
creating a code firewall here, we can limit recompilation times should
the other code be updated.