Merge pointer into underlying base type
This actually makes more sense. The `LocalResource` class is just
supposed to be a simple RAII wrapper class to handle automatic deletion.
It should be used as a drop-in replacement type, that functions
basically exactly the same as a raw pointer, just with automatic cleanup
in the event of exceptions. The template parameter type should match the
original type.

As the original code used `LPWSTR* commandLineArgs`, that type should be
maintained unmodified within the template type parameter. The weirdness
of peeling off a pointer in the type is not so obvious for
`commandLineArgs`, which is already a double pointer (a pointer to an
array of pointers to strings).  When working with a plain raw single
`LPWSTR` rather than an array, the difference becomes more obvious. How
do you specify the type `LPWSTR` with the pointer stripped off? There is
no `WSTR` type. By maintaining the full original type, it becomes easier
to work with both pointers to arrays of strings, and pointers to
strings.