Remove need to `unique_ptr` casts
A `unique_ptr<Derived>` can auto convert to a `unique_ptr<Base>`.

A reference to a `unique_ptr` does not auto convert based on the type of
pointer to object. To enable auto conversion, the `unique_ptr` must be
passed by value.

Passing by value requires an explicity `std::move` to mark the transfer
of ownership from an lvalue (roughly, a named variable). A transfer from
an rvalue (roughly, a temporary) does not need to be marked.

In parameter lists `unique_ptr` should always be passed by value.
Passing by reference defeats the ownership transfer semantics. It's not
unsafe to do, just kind of pointless. Either pass the `unique_ptr` by
value to transfer ownership, or pass the unwrapped underlying pointer
directly if the caller maintains ownership.