Lift `UNICODE` restriction on string conversion
Previously the code tried to issue a runtime exception for what was
effectively a compile time error. It also provided a misleading function
name in that the "T" in "LPCTSTR" meant it would adapt to changes in
string width based on the `UNICODE` setting. The function however only
actually provided support for narrow strings, not for wide strings.

By providing two overloads of the function, for both types of strings,
the compiler will auto detect at compile time which function should be
called. It also means both versions of the method could be used in the
same executable, allowing support of mixed string width code. If the
wrong string types are mixed together in an incompatible way, it will
result in a compile time error due to the type mismatch. And finally, it
means the function will finally live up to the "T" in its name.