Rework logging globals
The `LoggerFile` object is now constructed in DllMain.cpp, and is only
set as the logger after the start of `DllMain`. Any logged errors before
then will be ignored. In particular, if the `LoggerFile` instance
encounters an error trying to open its log file, it won't try to log it.

Visibility of the global `LoggerFile` instance has been reduced, as it
is now no longer declared as `extern` in any header file. Instead, it is
passed to the logging code and setup as the active logger. All log
output should go there to the global functions to be directed to the
logger.

The `logger` pointer is now found in Log.cpp, where it's logically
related to. It is defined and initialized such that it is guaranteed by
the C++ standard to be initialized before any globals in other
translation units that require dynamic initialization. This means the
`Log` function can expect it to have a known value, even before DllMain
has started. Further, the `Log` method can handle `nullptr` values, so
there is no risk of crash if `Log` is called before a logger is set.