Remove explicit `LoggerFile` destructor
As the `LoggerFile` is being destructed, all sub-objects will also be
destructed. The destructor of `std::ofstream` will already be
responsible for closing the associated file, so there is no need to be
explicit about this.

As per the `std::ofstream::close` reference documentation:
"This function is called by the destructor of basic_ofstream when the
stream object goes out of scope and is not usually invoked directly."
Source:
https://en.cppreference.com/w/cpp/io/basic_ofstream/close