Reorder `TrimOption` enum
This ordering allows for bit-field checks, which could simplify and
speed up the `if` checks. Example:
```
if (trimOption & TimeOption::Trailing) // Trailing or Both
```

Although ideally, when a parameter is likely to be passed as a fixed
constant, that's often a sign you should provide multiple functions
instead of a single function with an enum parameter.