Force non-unicode build of WindowsModule.cpp
After writing up details of the problem in Issue #148, I realized the
affected code was quite isolated from the rest of the program. Access
was controlled by a simple header file with just a function signature,
and none of the parameter or return types depended on the `UNICODE`
setting.

By simplying undefining the `UNICODE` and `_UNICODE` setting, this
module could be forced to build against the narrow string Windows API
interface. The setting changes will not escape to the rest of the
program due to the limited API of this file.