Use `inputWideString.size()` instead of NULL scan
If we have the size available, why not use it. It is efficiently cached
as part of the `std::wstring_view`.

Small change in behaviour here. The `WideCharToMultiByte` function will
include the null terminator in the byte count if it processes it. In
particular, if an input size of -1 is passed, the output size will
include space for the null byte.

The `size()` property of `std::wstring_view` (or even `std::wstring`)
does not include space for a terminating null. This means the
`WideCharToMultiByte` method will not automatically null terminate the
output, and the converted size will reflect that. The `std::string`
buffer output will already safely contain a terminating null, so it
doesn't need to worry about that, but the string clipping will need to
be adjusted for the byte count difference.