Convert `moduleDirectory` to function level static
This restricts access to the variable, and frees it from global
initialization order problems. Function level statics are initialized on
first use, and are not subject to the unsequenced initialization order
problems of globals.

As this variable was only ever being used from the `ConsoleModuleLoader`
class, it makes much more sense to tie it to the class. It was
previously declared global, but was not really used globally.