Use array type syntax at point of use
This is perhaps a bit more descriptive of what's actually being used. It
also makes more sense when the array subscript operator `[]` is used
later in the code.

The `unique_ptr` type only provides the `operator[]` member when the
underlying type is an array type. This allows for a closer match in
behaviour.