Use `NULL` in `dllHandle` check
The documentation for `LoadLibrary` uses the custom return type
`HINSTANCE`, and the documentation states that on error the function
returns `NULL`. Best to follow the documentation exactly here, in case
the return type definition ever changes.

Interestingly the compiler doesn't complain about the use of `NULL` here
when `-Wzero-as-null-pointer-constant` is used. This is despite `NULL`
being a macro for `0`, and its use generates warnings elsewhere. I
believe this is because of the `typedef` used for the return value. An
`HINSTANCE` is a `typedef` for `HANDLE`, which is a further `typedef`
which eventually works out to `void*`. The `HANDLE` type is generally
considered to be an opaque pointer. Perhaps they've somehow hidden the
internal `void*` details behind this opaque handle, which may be why
`NULL` doesn't generate a warning in this context.