Wrap raw string with `std::string` constructor
We don't really need a method to wrap raw strings. The constructors of
`std::string` and `std::wstring` do that already. We just need to know
the proper type. But if the type is already known, we can use it
directly. In particular, the code here uses a fixed parameter type, so
we will always have a fixed output type.

If the type is a bit more dynamic, say chosen with a `#define` setting,
such as `UNICODE`, which also sets the `TCHAR` value, then a type alias
is sufficient:
```
using tstring = std::basic_string<TCHAR>
```

If instead the type is chosen through a template parameter, then yes, we
do need some kind of wrapper to auto convert the type during template
instantiation.