Use native C++ types for `WrapRawString`
This makes the method usable without the `<windows.h>` header file being
seen.

This introduces the possibility of the Windows types being changed so
that they no longer match the declarations here. However, as we already
make assumptions about being able to map Windows types to `std::string`
or `std::wstring` internally, the code is already brittle in that
regards, and would need internal updates anyway.