Add `ConvertNarrowToWide` function
We don't really need this, but it does increase the symmetry of the API.

This method may potentially be useful for filename conversion, since the
"A" version of Windows methods don't support UTF-8. You would need to
convert to UTF-16 and use the "W" version of functions to open a file
with a non-ASCII / non-codepage character in the filename.