Convert `Whitespace` constant to `std::string_view`
Assigning a string literal such as `" \t"` to a `std::string` or a
`const std::string&` causes the `std::string` constructor to be called.
When a `std::string` parameter is defaulted in this manner, this
construction happens at every point of call where the default value is
used.

By using a `constexpr` value, the construction of the default parameter
can be moved away from the runtime point of calls to compile time.
However, the `std::string` type can not be initialized as `constexpr`,
as it does not consider `const std::string` to be a literal type.

A workaround is to use a `std::string_view`, which can be `constexpr`
initialized. It is also very cheap to copy, which means it is typically
passed by-value rather than by-reference. By using `std::string_view`
for the defaulted parameter, expensive constructor calls can be avoided
when calling methods with default values.

This result was confirmed using the Godbolt compiler.

Internally, the uses of `whitespace` have overloads which accept an
`std::string_view` parameter, so this works conveniently without
internal conversion either.