Convert `ConsoleModuleLoader` constructor to take a `vector`
This should complete the interface changes for the `ConsoleModuleLoader`
class.

Currently only the first element of the vector is actually processed. If
more than one module name is passed, an exception is thrown to signal
the unsupported case. That restriction will be removed once the
internals change to support the update.