Move comments from header to implementation file
This removes a duplicated comment. We should have comments in only one
place to reduce the risk of them getting out of sync. By putting them in
the implementation file, it's easier to verify the comment matches the
implementation. It also means updates to comments don't trigger rebuilds
of a large set of files.