Enable parallelism when building
The CircleCI environment should allow for up to 2 CPUs, so we should try
to make use of that. The newer makefile system does actually parallelize
quite well.

Running the unit tests with `check-testLib` is only a single task, so no
real parallelism there, though it doesn't hurt to enable it either.
Might as well be consistent.