Use explicity `std::string` type
The old code was a bit confusing as to what the type was, and actually
perhaps even misleading. The `ConsoleModuleLoader` constructor takes
`std::string` inputs, so that was probably the intended type. The type
deduction however produced a `char const* const`.

To declare a `std::string` use either of:
```
const std::string moduleName("NoDllTest");
const auto moduleName = std::string("NoDllTest");
```