Do not modify `loadOffset` on error
It shouldn't matter if `loadOffset` is changed or not, since it won't be
relevant if the Outpost2.exe module fails to load. Still, it's cleaner
to not modify it.

The real change though is having a code path that can mark when memory
patching is enabled versus disabled. That should match having
Outpost2.exe present in memory or not. It should perhaps also require
calling this method first to ensure relocations are properly handled.