Add `memoryPatchingEnabled` global
This value can be used to replace the `memoryCommandsDisabled` global.

By setting `memoryPatchingEnabled` after checking for the "Outpost2.exe"
module, we can more simply make patching of Outpost2.exe dependent on
having Outpost2.exe loaded into memory. This will be true when running
as part of the game, and false when running from a unit test executable.

If we later add more advanced unit testing that also loads Outpost2.exe,
it will naturally be true there once Outpost2.exe is loaded and the
`SetLoadOffset()` method is called is account for relocations.

For now, leave in `memoryCommandsDisabled`, as it removes the error
reporting when "Outpost2.exe" isn't found in memory.