Update check for empty console module name
Eventually we should use an empty vector for no console modules to load.
Specifying a module to load with an empty name will become an error.

For now, avoid throwing a real error when passed a single module name
which is an empty string. The old code used an empty module name to
represent no module to load. The interface hack that passes a vector of
1 element might now pass a vector of 1 element with an empty string,
rather than a vector of 0 elements.