Display error for empty console module names
Passing any empty names for console modules to load is an error.

Currently there is one exception due to the temporary check above. For
compatibility reasons, if only 1 module name is passed, it may be empty,
and the code will return without reporting an error.

The temporary check only applies when a single module name is passed. If
more than 1 module name is passed, they are all checked for empty
strings, which may result in an error.