Update pre-check before loading modules
We want to avoid calling `HookFileSearchPath()` if not needed. If no
modules are loaded, we don't need that hook.

At this point, anything in the `modules` vector will have a non-empty
module name. As such the check should be against the size of the
`modules` vector, rather than check if any of the module names within it
are empty.