Remove duplicate directory check
The `ConsoleModuleLoader` constructor already checks all module
directories to make sure they exist. We don't need to check it again
just before trying to load the DLLs. The DLL loading code is already
able to handle loading errors.

If it weren't for the more user friendly error messages, it would be
questionable to check the directory's existence at all. Even with the
check, we are still vulnerable to time-of-check to time-of-use race
conditions.