Cleanup test module folders using Win API
This works around a bug in the std::filesystem::remove implementation
provided by Mingw.

There is also a `rmdir()` POSIX method, but it is Linux only. There is a
`std::remove` standard library method, but that only works on files.
There doesn't appear to be a standard cross platform way to remove
directories in C++ before the `std::filesystem` API.