Add `static` version of `IniFile::GetValue`
This allows for quick one-off use, which avoids an explicit object
instantiation. This may be slower for repeated use to fetch multiple
values. In the current implementation it's not less efficient, but if
the implementation was changed to read and parse the ini file directly,
and store the contents in memory, it would be.

This is part of a refactoring to bring all INI reading code into a
single file.