Move `ResManager` to .cpp file
None of the method signatures or data types in the `ConsoleModuleLoader`
class depend on this class. That's true of both public and private
members. Effectively the `ResManager` class is an implementation detail
for the event hooking. The only dependence is in the reverse direction,
in that `ResManager` needs access to certain private members.