Add trailing slash after converting to string
The `fs::path::operator/()` has an odd behaviour, where if you try to
append an absolute path, it replaces the original path rather than
append to it.

On Linux a backslash (\) is not considered a path separator, but rather
just an ordinary character that can be part of a filename. That means
trying to append a backslash character can have different effects beween
the Windows and Linux platforms.

Adding the trailing slash after converting to a `std::string` avoids the
special behaviour.