Update dependency flags to include system folders
Although this produces larger dependency files, it is more accurate. In
particular, it will cause a rebuilt upon a compiler or system library
upgrade.

This shouldn't trigger extra rebuilds when switching between compilers
since the intermediate folders (and hence the dependency files) are
already split by compiler configuration.

Finally, this uses a flag set that is compatible with the NASM
dependency tracking flag set.