Add default value support to INI file
It was not possible to add default value support for both the static
method and the non-static `GetValue` method. Doing so would create
ambiguity between them. Static methods may be called on member objects,
and passing 3 parameters would make the call to which method ambiguous.
It might be an explicitly specified value for the last parameter to the
non-static method, or it might be a defaulted parameter call to the
static method. The C++ standard does not allow such overloads.