Remove faulty include
Source files (.cpp, .c) should never be included. Instead, the file
should be compiled directly. In fact, including the .c file from a .cpp
file defeats the purpose of it, as it would then be compiled as C++
code, and so would no longer be a valid check for C compatibility.