Add extra `EXPECT_CALL` to silence warnings
More long term, we should uncouple the `LogError` and `Log` methods.

The long term plan is to have a log distributor class that can
distribute a message to multiple outputs. That would allow a warning to
be logged by both a pop-up message box handler, and a log file handler,
and without direct internal calls within `LogError`.