Move `modulesRunning` lock point
Realistically things are fixed once the VOL list is fixed. That happens
at the call to `LoadVolFiles()`. Since there are no code hooks to run
module code between the end of `OnInit()` and the beginning of
`OnLoadShell()`, these are effectively the same point.

One small caveat, there is a possibility that a module installed its own
code hook, that could be triggered between the end of `OnInit()` and the
beginning of `OnLoadShell()`. However, as the VOL list has already been
fixed at this point, it would already be an error to modify it. It
simply wouldn't be detected and reported as an error.