Change pointer integer types to `uintptr_t`
Technically pointers to objects and maximum size of objects can be two
different types with two different sizes. It just so happens that in
today's typical flat memory models, the two types coincide. Previously,
on segmented memory models, this was not the case. For 16-bit x86 real
mode programming, object size might be restricted to a 16-bit segment,
while an address might include both a segment and an offset, resulting
in a 32-bit pointer.