Make `FindModuleHandle` visible externally
Because of the `HMODULE` return type, we now need to include
`<windows.h>` from the header file. This means any including files now
also sees the contents of the Windows header file. Fortunately the only
include of this header is from op2ext.cpp which already includes the
Windows header file.